Header file modifications
FLAME.h and other header files that it depends on, have declarations
and definitions that are conflicting with BLIS header file, blis.h.
This creates issues when an application includes both FLAME.h and
blis.h. This commit modifies and cleanup of the existing header files
to address this issue.

AMD-Internal: CPUPL-2898
Change-Id: I6ec8b1ed5c75ce4fd9f1955b3fc1d857201dac20