DGELS optimization for small size
Optimize DGELS for small sizes between 8 to 100.
Optimizations steps
- DLANGE API vectorized for functionality that finds largest number
in a matrix
- Skip further processing if last column is found to be 0 on entry
in DLARF
- Vectorized code in DLARF fusing DGEMV and DGER operations in the
path taken for left apply of elementary reflector. AVX2 in this
commit. AVX512 will be done in followup commit
- Add ctest for covering performance tests for DGELS with M in range
10-40 and N 10

Few Eigen netlib tests fail threshold test with the current
optimization. DEV and DVX netlib tests compare Eigen values generated
by DGEEV call where only Eigen value is requested vs when both Eigen
values and vectors are requested. But the test compares the output
directly by value and not norm of their difference. Hence those
failures are ignored as the difference in output with reference and
intrinsic path is in 15th decimal digit or higher.

AMD-Internal: CPUPL-4593
Change-Id: I866de586a4f57788704c75b862361efa730c940f