Added  singularity check in SPFFRT2 and SPFFRTX routines
SPFFRT2 and SPFFRTX routines are designed for partial factorization and does not perform pivoting. Currently, When a zero diagonal element is encountered, the routines does not handle it explicitly, leading to numerical issues such as division by zero or infinite values. To address this, we have added a check that skip trailing matrix update whenever a zero diagonal element is encountered.

Though this change improves routines stability, we will not be able to accurately reconstruct the original input matrix from the API output

Change-Id: If80ffcea1b879bd8697d03454dc19828cb228d54
Signed-off-by: Sridhar Govindaswamy <Sridhar.Govindaswamy@amd.com>